feat: 투입 금액의 잔돈 출력하기 - 기존 클래스명 변경 및 새로운 클래스 추가   - Change: 투임 금액의 잔돈을 저장   - Inventory: 재고 상품을 관리
