+++ src/main/java/vendingmachine/Application.java
@@ -1,23 +1,23 @@
 package vendingmachine;
 
-import java.util.Arrays;
-
 public class Application {
     public static void main(String[] args) {
         Message message = new Message();
         User user = new User();
-        Change change = new Change();
+        Inventory inventory = new Inventory();
         Products products = new Products();
 
         message.printInputHolding();
         int holding = user.inputHolding();
-        change.makeCoins(holding);
+        inventory.makeCoins(holding);
 
         message.printInputProducts();
         String[] productsList = user.inputProducts();
         products.addProducts(productsList);
 
         message.printInputAmount();
-        user.inputAmount();
+        int amount = user.inputAmount();
+        Change change = new Change(amount);
+        message.printRemainingAmount(change.getAmount());
     }
 }
