+++ src/main/java/vendingmachine/User.java
@@ -25,11 +25,11 @@ public class User {
         }
     }
 
-    public String inputAmount() {
+    public int inputAmount() {
         while (true) {
             String input = Console.readLine();
             if (validation.isValidateAmount(input)) {
-                return input;
+                return Integer.parseInt(input);
             }
         }
     }
